SDC-9066. Refactor Avro to Parquet Convert Validation and ParquetWriter
Refactored the following from mapreduce-protolib to wholefile-convert-protolib
* parquet constants
* validation and parquet writer

Reviewed-on: https://review.streamsets.net/14694
Tested-by: StreamSets CI <streamsets-ci-spam@streamsets.com>
