+++ mapreduce-protolib/src/main/java/com/streamsets/pipeline/stage/destination/mapreduce/config/JobConfig.java
@@ -25,7 +25,7 @@ import com.streamsets.pipeline.stage.destination.mapreduce.Groups;
 import com.streamsets.pipeline.stage.destination.mapreduce.MapReduceErrors;
 import com.streamsets.pipeline.stage.destination.mapreduce.jobtype.avroconvert.AvroConversionCommonConfig;
 import com.streamsets.pipeline.stage.destination.mapreduce.jobtype.avroorc.AvroOrcConfig;
-import com.streamsets.pipeline.stage.destination.mapreduce.jobtype.avroparquet.AvroParquetConfig;
+import com.streamsets.pipeline.lib.converter.AvroParquetConfig;
 import org.apache.hadoop.mapreduce.Job;
 
 import java.util.ArrayList;
