+++ mapreduce-protolib/src/test/java/com/streamsets/pipeline/stage/destination/jobtype/avroparquet/AvroParquetConvertIT.java
@@ -25,8 +25,8 @@ import com.streamsets.pipeline.sdk.ExecutorRunner;
 import com.streamsets.pipeline.stage.destination.mapreduce.MapReduceDExecutor;
 import com.streamsets.pipeline.stage.destination.mapreduce.MapReduceExecutor;
 import com.streamsets.pipeline.stage.destination.mapreduce.jobtype.avroconvert.AvroConversionCommonConfig;
-import com.streamsets.pipeline.stage.destination.mapreduce.jobtype.avroparquet.AvroParquetConfig;
-import com.streamsets.pipeline.stage.destination.mapreduce.jobtype.avroparquet.AvroParquetConstants;
+import com.streamsets.pipeline.lib.converter.AvroParquetConfig;
+import com.streamsets.pipeline.lib.converter.AvroParquetConstants;
 import org.apache.avro.Schema;
 import org.apache.avro.util.Utf8;
 import org.apache.hadoop.fs.Path;
