+++ wholefile-converter-protolib/src/main/java/com/streamsets/pipeline/lib/converter/AvroParquetConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.streamsets.pipeline.stage.destination.mapreduce.jobtype.avroparquet;
+package com.streamsets.pipeline.lib.converter;
 
 import com.streamsets.pipeline.api.ConfigDef;
